refactor(assistant-service): Refactored assistant-service to use a domain-based project structure
