+++ assistant-service/app/core/auth.py
