+++ assistant-service/app/core/crypto.py
