+++ assistant-service/app/core/database.py
