+++ assistant-service/app/llm/schemas.py
