--- assistant-service/app/services/__init__.py
@@ -1,1 +1,0 @@
-from .llm_service import get_llm_client
